# This is a combination of 2 commits. # This is the 1st commit message:
Coordinate: Add std::pair constructor

# This is the commit message #2:

Fix std::pair member types

+++ src/Coordinate.cpp
@@ -40,7 +40,7 @@ Coordinate::Coordinate() : Coordinate::Coordinate(0, 0) {};
 Coordinate::Coordinate(double x, double y) : X(x), Y(y) {};
 
 // Constructor which accepts a std::pair for (X, Y)
-Coordinate::Coordinate(const std::pair<int, int>& co)
+Coordinate::Coordinate(const std::pair<double, double>& co)
 	: X(co.first), Y(co.second) {};
 
 // Generate JSON string of this object
